third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/querySelector-with-detached-node-distribution-expected.txt

CONSOLE ERROR: line 87: Uncaught TypeError: element.createShadowRoot is not a function
crbug.com/337618: querySelector needs ensure distribution.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
